Separate helpers geojson.mjs to each functions (backport from geotoolbox)
